+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## 2.3.6
 
-* Adds setMaxVideoResolution method to support different video resolutions while picking a video stream from a source with different tracks
+* Adds setMaxVideoResolution method to support different video resolutions while picking a video stream from a source with different tracks.
 
 ## 2.3.5